fuzzy finder: Refactor
The main purpose of this commit is to configure the fuzzy finder
keyboard shortcut to also be triggered when the target is an input
element. This allows us to remove `onHandleFuzzyFinder` from the query
input, making the component APIs simpler.

Since we don't have pass down that handler anymore we can now also
consolidate the fuzzy finder logic into its own component. I considered
just moving everying into `FuzzyFinder` itself, but to keep changes of
existing code to a minimum I decided to introduce `FuzzyFinderContainer`
instead.
