+++ client/search-ui/src/input/CodeMirrorQueryInput.tsx
@@ -76,7 +76,6 @@ export const CodeMirrorMonacoFacade: React.FunctionComponent<React.PropsWithChil
     onBlur,
     isSourcegraphDotCom,
     globbing,
-    onHandleFuzzyFinder,
     onEditorCreated,
     interpretComments,
     isLightTheme,
@@ -212,10 +211,9 @@ export const CodeMirrorMonacoFacade: React.FunctionComponent<React.PropsWithChil
                 onFocus,
                 onBlur,
                 onCompletionItemSelected,
-                onHandleFuzzyFinder,
             })
         }
-    }, [editor, onChange, onSubmit, onFocus, onBlur, onCompletionItemSelected, onHandleFuzzyFinder])
+    }, [editor, onChange, onSubmit, onFocus, onBlur, onCompletionItemSelected])
 
     // Always focus the editor on 'selectedSearchContextSpec' change
     useEffect(() => {
@@ -435,10 +433,7 @@ export const CodeMirrorQueryInput: React.FunctionComponent<
 // Instead of creating a separate field for every handler, all handlers are set
 // via a single field to keep complexity manageable.
 const [callbacksField, setCallbacks] = createUpdateableField<
-    Pick<
-        MonacoQueryInputProps,
-        'onChange' | 'onSubmit' | 'onFocus' | 'onBlur' | 'onCompletionItemSelected' | 'onHandleFuzzyFinder'
-    >
+    Pick<MonacoQueryInputProps, 'onChange' | 'onSubmit' | 'onFocus' | 'onBlur' | 'onCompletionItemSelected'>
 >({ onChange: () => {} }, callbacks => [
     Prec.high(
         keymap.of([
@@ -457,19 +452,6 @@ const [callbacksField, setCallbacks] = createUpdateableField<
             },
         ])
     ),
-    keymap.of([
-        {
-            key: 'Mod-k',
-            run: view => {
-                const { onHandleFuzzyFinder } = view.state.field(callbacks)
-                if (onHandleFuzzyFinder) {
-                    onHandleFuzzyFinder(true)
-                    return true
-                }
-                return false
-            },
-        },
-    ]),
     EditorView.updateListener.of((update: ViewUpdate) => {
         const { state, view } = update
         const { onChange, onFocus, onBlur, onCompletionItemSelected } = state.field(callbacks)
